[bugfix] Fix macro guard in overlappingschwarztest
The loaction of the macro guard lead to the method `test`
not returning a value if neiter SUperLU nore SuiteSparse
are found. While the test is not executed in this case due
to a CMake guard, these macro guards should be placed
in the correct place anyways for reasons of code hygiene.