Move compressed-column matrix into a separate file
The file colcompmatrix.hh contains the rudimentary implementation
of a sparse matrix with scalar entries and column-compressed storage.
As part of cleaning up the UMFPackSolver implementation this patch
moves the matrix class into a separate file.  What's left in
colcompmatrix.hh is the matrix setup code.

The plan is to turn the matrix into something ressembling a real
dune-istl matrix.  That should make the code easier to understand,
and maybe something will even need compressed-column matrix
for other things one day!