BCCSMatrix: Remove assignment from matrix of different types
Remove operator= and setMatrix methods.  The code of these methods
is copied to all locations were operator= or setMatrix were
previously used.  This makes the calling code a bit longer,
but it simplifies the overall code structure considerably.

In particular, with these changes the BCCRSMatrix class does not
have 'virtual' methods anymore.