BCCSMatrix: Make first template parameter the entry type
Previously, the first template parameter was the matrix type that
the BCCSMatrix was a copy of.  Weird, and very closely tied to
the fact that BCCSMatrix/ColCompMatrix is currently only used
to get matrix data into SuiteSparse solvers.

This patch makes the first template parameter the type of
matrix entries.  This is what all other dune-istl matrices do,
and hence the code again gets easier to understand.