Implement generic vector iterator
The class `VectorIterator` provides the method `iterate( entryAction )`.
The `entryAction` is a lambda which is called at each (scalar) entry of the vector.
The lambda takes the following arguments: index as `std::tuple<...>` and the entry itself.
`std::tuple<...>` is chosen to address static multitype vector types.

Currently, arbitrary nested combinations of `FieldVector`, `BlockVector` and `MultiTypeBlockVector` are possible.