UMFPACK: Use generic flatMatrixForEach routines
Instead of calling the BCCSInitializer, we directly parse the matrix into
an UMFPACK compatible form. This allows us to use almost every blocked
matrix structure, such as MultiTypeBlockedMatrix. Moreover, the 'setMatrix'
method is extended by a second bitVector argument to handle exclusion of
indices. The old 'setSubMatrix' method is kept. The new interface is
inspired by the neighbored 'Cholmod' class.

The additional routines on the vector data are linear in time and have no
measurable influence on the run time of UMFPACK.