Speed up MatrixIndestSet::exportIdx()
Before this patch `BCRSMatrix::addindex()` was used on individual
column indices. This is slow because each insertion does a binary
search although the inserted indices are already sorted. Bulk-inserting
whole rows with `setIndices()` improves on this significantly but still
does a non-necessary sort. The latter is avoided by the new
`BCRSMatrix::setIndicesNoSort()` method.