Export missing functionality in MatrixIndexSet
The `MatrixIndexSet` class provides access to an
incomplete subset of its internal data. While
the number of rows is exported, the number of cols
and the column indices per row have been missing
so far. This was an oversight, since one can hardly
use the class without this, unless one relies on
the dedicated `exportIdx()` method for `BCRSMatrix`.

With the missing data exported, one can now easily
use `MatrixIndexSet` to build patterns for
other sparse matrix classes.