[solverfactory] Defer UMFPack check, as non-double field type triggers a compile error.
In the rewrite the vector types of UMPack were checked, similar to the
other direct solvers. This does not work anymore, as recent changes in
UMFPack now lead to a compiler error if the vectors are not available,
i.e. when using a real-type different from double. This change now
(re-)introduces a SFINAE check to ensure that the types are compatible.