Fix memory leak in the CHOLMOD wrapper code
When calling CHOLMOD to compute the Cholesky factor of a matrix,
it was not tested whether such a factor already existed. As a
consequence, if the solver was called twice in a row, the old
factor would be lost to a memory leak.

Fix this by explicitly deallocating the old Cholesky factor
before asking CHOLMOD for a new one.

(cherry picked from commit e4ffbb1f1540326817ee080231d6f699efa3c4b5)