Reuse RowIndexSet in BCRSMatrix::CreateIterator
The BCRSMatrix with row_wise creation makes many allocations and uses very sparse memory. This leads to a relatively slow pattern creation. So this change exposes the internals of the (faster) row index set from the MatrixIndexSet class into an re-usable pattern creator for the row-wise build mode of the matrix.